predictor/requirements.txt
astroid==1.6.0
Django==2.0.1
django-cors-headers==2.1.0
djangorestframework==3.7.7
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.13.3
pandas==0.22.0
pylint==1.8.1
python-dateutil==2.6.1
pytz==2017.3
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
wrapt==1.10.11
